moved functions to component level
